+++ model/db/Project.cs
@@ -11,10 +11,10 @@ namespace btl_web_nangcao_task_management_system.model.db
         {
         }
 
-        public int id { get; set; }
+        public long id { get; set; }
         public string title { get; set; }
         public string description { get; set; }
-        public int lead { get; set; }
+        public long lead { get; set; }
         public DateTime startDate { get; set; }
         public DateTime estimateDate { get; set; }
         public ProjectStatus status { get; set; }
